Crate miau

miau is a community IRC bot with a continuous delivery workflow allowing participants to have their changes deployed to the running instance within minutes.

Modules

bot
The main bot entry point.

commands

environment
Environment and configuration management.

irc
IRC handling functions

logging
Basic logger implementation that prints to standard output.

network
IRC network state management and protocol handling.