Module miau::environment

Environment and configuration management.

"Environment" is just another word for configuration, but the idea is that everything that could affect the way the bot makes configuration-based decisions should be encapsulated in the structures found in this class, most notably Env.

Since the environment must be loaded before anything else can happen, the load takes no arguments. Anything that could affect the way the environment is loaded should itself be part of the environment, so the logic for detecting those things should live in this module.

Structs

EnvInner
The main struct for accessing the bot's environment, including configuration loaded from files, the command line, and environment variables. Any static user configuration should be retrieved from an instance of this struct.

Enums

Error
Used to signal that there was an error loading the environment.

Functions

load
Loads the bot's environment.

lookup
Look up a value. Splits on periods, descends that way.

Type Definitions

Env