Struct miau::environment::EnvInner

pub struct EnvInner { /* fields omitted */ }

The main struct for accessing the bot's environment, including configuration loaded from files, the command line, and environment variables. Any static user configuration should be retrieved from an instance of this struct.

Methods

impl EnvInner

fn conf<'a>(&'a self, path: &'a str) -> Option<&Value>

Fetches the given configuration value, if it exists. Refer to the lookup documentation for the meaning of the "path" argument.

fn conf_str<'a>(&'a self, path: &'a str) -> Option<&str>

Fetches the given configuration value as a string slice, if it exists.

fn conf_integer<'a>(&'a self, path: &'a str) -> Option<i64>

Fetches the given configuration value as an integer, if it exists.

fn conf_integer_or<'a>(&'a self, path: &'a str, or: i64) -> i64

Fetches the given configuration value as an integer, or the default value, if it doesn't exist, and prints a warning.

fn conf_float<'a>(&'a self, path: &'a str) -> Option<f64>

Fetches the given configuration value as a floating point value, if it exists.

fn conf_bool<'a>(&'a self, path: &'a str) -> Option<bool>

Fetches the given configuration value as a boolean, if it exists.

fn conf_array<'a>(&'a self, path: &'a str) -> Option<&Vec<Value>>

Fetches the given configuration value as an array of TOML values, if it exists.

fn conf_table<'a>(&'a self, path: &'a str) -> Option<&Table>

Fetches the given configuration value as a TOML table, if it exists.